Update patch set 5
Patch Set 5:

I don't know. It is very stupid. I changed the name to "barometer_whatever", install the module again, and it works.. I move back to "barometer" and it fails... There is no other barometer module installed on the system... very weird..

If you don't like that long name we can change it by something shorter, but barometer should be part of the module name.

Patch-set: 5