Update patch set 5
Patch Set 5: Code-Review+1

I totally agree that 'barometer' name is ideal (most clear).  Did it work yesterday when we first did these patches?  Could something be stale in the system such that now sees it as a conflict/duplicate module?!  ...Maybe alternatively we can shorten it to 'baro' to keep it more neat and avoid lengthening it, but it's just a suggestion.

Patch-set: 5
Label: Code-Review=+1