Update patch set 4
Patch Set 4:

I would have considered here to simply push the file hosted in Functest. Then we could update our testcases.yaml to refer to the new package.

Then in a second change, the code related to functest in collectd could have been moved to the subclass of TestCase. It seems a better design.

Patch-set: 4