Update patch set 1
Patch Set 1:

(1 comment)

@Taras, you need to add the license header

@Al, if I'm not mistaken, the  RPMs are created when you build collectd, and depends on what plugins are enabled.
The enabled plugins depend on what libraries are installed on the system.
It is possible that OS updates, collectd updates, and updates to the build node would result in additional libraries being enabled, and additional plugins being enabled, therefore generating additional RPMs. 
So, counting may not be sufficient, and may produce a false positive in the case where an additional RPM is generated by collectd, and one of our RPMs is not built.

Patch-set: 1