Update patch set 1
Patch Set 1: Code-Review+2

This change is completed as per Maryam's specifications.
We can add a new change after the release to introduce a new efficiency. I'd like to see this change in for the E release, so that there's a little more certainty, and we can detect issues quicker.

Patch-set: 1
Label: Code-Review=+2