Update patch set 1
Patch Set 1:

(5 comments)

Looks mostly fine to me, content wise. I've left a few comments inline.

Main takeaway: the markup is wrong if you're looking for monospaced font.
I've highlighted SOME occurrences in the doc, so if you're going to change it, that's not an exhaustive list.

Patch-set: 1
Label: Verified=0