Update patch set 6
Patch Set 6:

Actually, this build was successful. 

It's the next patch set in the series that changes the repo address[1], but that can't merge until the repo is renamed by OpenStack infra. There is also a corresponding change in Apex which is dependant on the repo url changing[3].

[1] https://gerrit.opnfv.org/gerrit/#/c/41393
[2] https://review.openstack.org/#/c/500768/
[3] https://gerrit.opnfv.org/gerrit/#/c/41405/

Patch-set: 6
Label: Verified=0