VES: add new tag to strip extra dashes
The `!StripExtraDash` tag can be used in YAML configuration
file to strip extra dashes in the YAML filed value. This may
happen if some of the collectd plugin field (plugin_instance,
type_instace etc.) is set to empty value but the field is
specified in the YAML configuration file. For instance:

"{vl.plugin_instance}-{vl.type_instance}"

The result fo the example above can be string consisting
extra dashes if `type_instace` field is not set by a collectd
plugin:

"load-"

Change-Id: I5cf7b19902acdfb0b4a93f3bf0c8b8e05cff2e97
Signed-off-by: Mytnyk, Volodymyr <volodymyrx.mytnyk@intel.com>