Update patch set 2
Patch Set 2:

> do we really need all the conf files here? the docker image will
 > have them all already. All ansible needs to do is disable the ones
 > that aren't relevant for the platform.

Or do you want to override the ones in the docker container? I still think that we are better off to simply remove the ones we don't want from inside the docker container - or we override with ones from the barometer repo? 
OR is this for offline installation?

Patch-set: 2