Fix syntax error in baro_test/collectd.py:mcelog_install
The version of CentOS that's used no longer matches the version that
is explicitly checked in collectd.py:mcelog_install.

As a result the code on L637-L642[1] is now being called.
This code was syntactically incorrect, however, was previously not being
called, witht eh update to the version, it is now being executed and
causing failures in deployment [2].

[1] https://github.com/opnfv/barometer/blob/c10400426df872be3236677dcb749c755f4a559c/baro_tests/collectd.py#L637-L642
[2] https://build.opnfv.org/ci/job/functest-apex-baremetal-daily-master/809/consoleText

Change-Id: I35274d4ede17510c3ae03c756eda7bc9f78b4e17
Signed-off-by: Emma Foley <emma.l.foley@intel.com>