Use multi-stage builds to reduce image size
There isn't a need to have all of the development dependencies
included in the distributed container image, so we can greatly reduce
the resulting distributed image through the use of multi-stage builds.

In the multi-stage build, we make sure of an interim image that builds
the artifacts we need, and then our final image can pull artifacts
from the interim image, thereby only pulling in the built artifacts,
and not the entire development environment.

For the first stage, the artifacts are built with the ci/ directory
scripts that generates the RPM files. In the second stage, those RPM
files are copied in and installed, thus avoiding shipping all the
development dependencies.

In my testing, this approach has reduced the resulting image from
1.72GB down to ~600MB.

Signed-off-by: Leif Madsen <lmadsen@redhat.com>
Change-Id: I4aa91bb4366c68e9873ba582d04693390ff744ca