Update patch set 7
Patch Set 7:

These changes will require an update to the releng/jjb/releng/opnfv-docker.yml file so that the build is executed from the root directory, pointing at the Dockerfile as a path. That's because I'm making use of the COPY command that will pull in the local files rather than cloning the repository.

Patch-set: 7