Update patch set 9
Patch Set 9:

I'm hoping this is ready for final review now. I've been using a container this way, and it seems to work well. The only other change I might suggest would be moving away from the entrypoint being a script, and moving it to the command directly so it's easier to override arguments to it.

Patch-set: 9
Label: Verified=0