Configure collectd to load ipmi before PMU
Collectd loads config files in alphabetical order.
Ansible creates the config files for collectd.
In this patch we change the name ipmi.conf to 0_ipmi.conf
This will allow ipmi to be loaded before PMU.
If PMU is loaded before ipmi the container will crash.
This is a bug in collectd 5.8

Change-Id: I3334a3a1970f596e36f665698649d6c04fd38fbb
Signed-off-by: John O Loughlin <john.oloughlin@intel.com>