Update patch set 3
Patch Set 3:

Hi @John, I've investigated this kind of crash some time ago. I've sent you an email with additional explanations.
In short the problem was with uuid plugin and not with ipmi or intel_pmu. For now I think it would be better to disable uuid plugin or patch it before build. Changing the order of plugins just hide the real problem and may cause collectd to be unstable.
Error has been already fixed by collectd community, but it will be available in next release (5.8.1? I don't know when).

Patch-set: 3