Update patch set 2
Patch Set 2:

@Emma, I added some comment about default behavior for empty string.
It is short and as for other plugins there is a reference to collectd documentation with more information.
A also added some example to sample configs, there is no specific deployment code for this option. For now I do not change the ansible scripts as the option is not available in current container version.

Patch-set: 2
Reviewer: Gerrit User 6511 <6511@bbac25d2-bf60-4904-9ba8-a72fc000d6c5>
Label: Verified=0