Update patch set 1
Patch Set 1:

> This will also catch res[3] == 0. Is that okay?

I think "not res[3]" is a safe choice...

Your comment seems correct to me.
The variable res[3] seems to be only True or False.
If so, your proposal is okay, because False is 0.

However, I can not understand why None was included in the expression. If it has some meaning, we can not catch with res[3] == 0, because None is not 0.

Patch-set: 1
Label: Verified=0