Update patch set 1
Patch Set 1:

Hi Aaron,

I did quick investigation and it seems that this error is not related to changes introduced in my commit.
The problem is in collectd code itself (in autoconf) and it is visible on this particular CI machine. During configure I can see that dpdkevents have been found but in old version (probably dpdk was installed during one of previous verification jobs - my particular CI job was not updating dpdk libs - it is only being done on demand when special configuration option is passed)
  dpdkevents. . . . . . no (DPDK version < 16.07)
  dpdkstat  . . . . . . yes

I think that some dependency is missing in autoconf - dpdk utils should not be built when there is no dpdkevents module in the system.
At least when I've compared logs to scenario that have both dpdkevents and dpdkstat in correct versions, similar collectd build ended successfully.
e.g. here is the link to verification of related patch from Kamil Wiatrowski that ended successfully:
https://build.opnfv.org/ci/job/barometer-verify-master/651/console

Patch-set: 1