Update patch set 1
Patch Set 1:

ok, thanks for review - it makes sense to simplify things that way:) in next iteration I'll refactor that code. So there would be param-setting only for insecure url because it needs to be added to whitelist before can be used in docker. And handling for secure urls will be much more cleaner.

Patch-set: 1