Update patch set 10
Patch Set 10:

This should be backported to stable/gambia, as it is affecting Yardstick/NSB when they try to use Barometer to collect NFVI metrics (i.e. deploying Barometer only and configuring it to report metrics to an existing influxDB instance).


Without this change, barometer deploys with 1 instance of InfluxDB; With this change, there is the flexibility to use/report metrics to 0 or more instances of InfluxDB.

Patch-set: 10