[docs][install] Split the docker and ansible install guides
docker.userguide.rst contained two methods of installation:

* Manual installation of all the docker containers
* one-click install using ansible

A new user is potentially going to be confused about which
document to use, and may not be aware that the two options exist.

This change splits the install guide into two documents, one for
docker, and one for ansible. It also renames the docker guide to
make it more obvious to the user what is contained in the guide.

Change-Id: I1ecdd7d15f8f53c80221916379982bca2673a06d
JIRA: BAROMETER-128
Signed-off-by: Emma Foley <efoley@redhat.com>