Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Thank you. Would it be possible to re-use the config file, so that you don't need to duplicate the file?

It is in line with configs for other plugins. One is just an example for manual use and other one is for ansible one-click install.
For stable container there are sample configs in 'src/collectd/collectd_sample_configs/' and in 'docker/ansible/roles/config_files/templates/' for ansible scripts.

Patch-set: 1