Update patch set 2
Patch Set 2:

> Patch Set 2: Code-Review-1
> 
> (1 comment)

I would like to differ with this comment. The current installation is written to work for a specific version set of components. It isnt part of the requirement to make the scripts compatible with multiple version of software, DPDK, Centos, Ubuntu or otherwise. We can add the requirement of supporting multiple versions of DPDK for J release. 
So for Iruya, we would only support specific version of DPDK & rest of the software.

Patch-set: 2
Reviewer: Gerrit User 1171 <1171@bbac25d2-bf60-4904-9ba8-a72fc000d6c5>
Label: Verified=0