Update SNMP agent config
The one-click install was not working with SNMP. The manual install
also wasn't working due to the ``Instance true`` config option in
the config. ``Instance true`` is not a valid config option for snmp
agent.
This option appears in the stable container config, and also in the
ansible config, but not in the master container. Removing this
unsupported option allows the config to be read and collectd to be
started.

JIRA: BAROMETER-134
Signed-off-by: Emma Foley <efoley@redhat.com>
Change-Id: If154b3190d82d1cd3638a9c258d2b48bb0efe9af