Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Thanks Emma for doing this. 
> Would it be better if we name it to barometer-collectd-main to better reflect collectd's main branch?

I considered that initially. I thought that "latest" would better reflect what we put in that container. It might cause confusion about what the barometer-collectd container is for if there's another one called barometer-collectd-main (since the stable one is our "main" container for users, with -master being for development and testing).
I agree that using -latest might also cause confusion, but I think it would be less, as -latest is IMHO more self explanatory.

I'm open to input on this, and thought that there might be some alternative suggestions, which is why I have a second change for the container flavour separate from this one. This master->main can merge immediately, and keep things up to date, but the -master -> -latest might change.

Patch-set: 1