[container][latest] Rename and update barometer-collectd-latest
Collectd changed the upstream default to main, and Barometer added a few
changes to incorporate this. One change was renaming the
`barometer-collectd-master` container to `barometer-collectd-latest`, since
`barometer-collectd-main` didn't make sense semantically, as we use the
stable container as the "main" reference.

At the time, we didn't have the changes made in releng to build the new
container, so container renaming was paused.
Now that the change is merged in releng (and has been for a while), and
the containers are being updated, this change should happen, so that the
container gets published.

Signed-off-by: Emma Foley <efoley@redhat.com>
Change-Id: I4ee91c289b9e7477157bea148452be58343dc733