[ci] Update DPDK package version
The nominal package for DPDK (19.11-3) is no longer available in CentOS
8. This is causing the container builds to fail [1][2][3]
The version has been updated to 20.11.

This may cause issues with the dpdkstat and dpdkevents plugins, however,
the dpdk_telemetry plugin is preferred over these.

The V1 legacy interface, accessed by dpdk_telemetry plugin, is also
exposed by DPDK 20.11. So the version can be updated to 20.11.

[1] https://build.opnfv.org/ci/view/barometer/job/barometer-collectd-docker-build-push-master/123/console
[2] https://build.opnfv.org/ci/view/barometer/job/barometer-collectd-latest-docker-build-push-master/17/console
[3] https://build.opnfv.org/ci/view/barometer/job/barometer-collectd-latest-docker-build-push-master/17/console

Change-Id: I33690a9694467a39b4f02f239197aa7cae064716
Signed-off-by: Emma Foley <efoley@redhat.com>