[ansible] Use collectd_config for virt and rdt
Only one of virt and intel_rdt can be enabled at a time, and the
existing checks are based on the existance of a {rdt,virt}.conf file to
determine whether the other plugin is enabled.
This check needed updating, and is now based on whether the other plugin
is in the collectd_plugins value.

Signed-off-by: Emma Foley <efoley@redhat.com>
Change-Id: I175d60111bf133965e1ecbf3764cb12a25356576