Update patch set 8
Patch Set 8:

> Patch Set 8:
> 
> @Emma, thank you for pushing next patch.
> I do not want to complain again, but it still does not install deps correctly:
> # dnf --showduplicate list collectd
> Last metadata expiration check: 0:33:29 ago on Mon 28 Jun 2021 11:57:46 AM UTC.
> Available Packages
> collectd.x86_64     5.10.0-1.el8      centos-opstools
> collectd.x86_64     5.10.0-3.el8      centos-opstools
> collectd.x86_64     5.11.0-1.el8      centos-opstools
> collectd.x86_64     5.11.0-2.el8      centos-opstools
> 
> # dnf builddep --enablerepo centos-opstools collectd
> enabling appstream-source repository
> enabling baseos-source repository
> enabling extras-source repository
> Last metadata expiration check: 0:32:11 ago on Mon 28 Jun 2021 11:59:08 AM UTC.
> no package matched: collectd
> Error: Some packages could not be found.
> 
> One solution might be to use spec file directly from opstools repo as bellow:
> # dnf builddep https://raw.githubusercontent.com/centos-opstools/collectd/master/collectd.spec
> But this would make dependency between collectd 5.12 and the latest spec file from repo, so I'm not sure if this is something we want.

There is a difference between centos-opstools and collectd/contrib/redhat/collectd.spec, however, we are trying use builddep to install the deps from the opstools repo anyway, so this makes sense.

I was thinking of using the patched spec file from the collectd repo, however, that would involve a refactor of the scripts in barometer/ci/, so I was pondering on how that could be achieved.
I think that your idea should work, and involve very little change.
I'll respin the patch, and hopefully merge it and the next in the series today, so I can see if the daily build job [1] is working again.

[1] https://build.opnfv.org/ci/view/barometer/job/barometer-daily-master/

Patch-set: 8
Reviewer: Gerrit User 2311 <2311@bbac25d2-bf60-4904-9ba8-a72fc000d6c5>