[docs] Add reno to help in creating release notes
Reno[1] is used to create release notes. It can be used in conjunction
with sphinx to automatically create and publish release notes. In this
case it is being added to help with release notes collation at the end of
the release cycle.

There is a possibility that it can be more fully integrated later, if
there is need for it.

For the moment, release notes can be outputted using tox::

  tox -e docs-releasenotes

Release notes can be added to features using::

  reno new slug-goes-here

where slug-goes-here is combined with a random suffic to create a unique
name for the release note.

[1] https://docs.openstack.org/reno/latest/

Change-Id: Id7f113fa4817585bb429503ff3491c633a4ce8e2
Signed-off-by: Emma Foley <efoley@redhat.com>