[docs] Update config and post-installation guides
The configguide was removed previously because it only contained
information about using the OPNFV installers, however, now there appears
to be no config information about the plugins. The config information
for each plugin is included in the user guide in
docs/release/userguide/feature.userguide.rst.
This change adds in a note to the the config guide on where to look.

Signed-off-by: Emma Foley <efoley@redhat.com>
Change-Id: I7f71d51846c9116bed3882b96e762ab5d70fdfa4