[docs][reno] Update reno to include template
The template for release notes wasn't conforming to the expected
headings. A template was added to the config.

To generate a new note (from the base directory in the repo), use::

    reno --rel-notes-dir docs/release/release-notes/ new slug-goes-here

The release notes are also included as part of the docs build, and the
existing releases notes have been renamed accordingly.

Signed-off-by: Emma Foley <efoley@redhat.com>
Change-Id: Id0cd7e00db35322f7b84cb06ebf8fe2a675d9941