[docs][reno] Update reno to include template
The template for release notes wasn't conforming to the expected
headings. A template was added to the config.

To generate a new note (from the base directory in the repo), use::

    reno --rel-notes-dir docs/release/release-notes/ new slug-goes-here

The release notes are also included as part of the docs build, and the
existing releases notes have been renamed accordingly.

Signed-off-by: Emma Foley <efoley@redhat.com>
Change-Id: I05d1950b37f266bb6eb3515b3660665cfd5c1d8b