[grafana] Update Dashboard for CPU Usage
After the introduction of the collectd_config roles, for generating
configs, some of the configs changed, as they were no longer relying on
collectd's defaults.
One of these was the cpu plugin, which configured ReportPercentage True
by the config role. The effected the grafana dashboards, because the CPU
usage dashboard explicitly selected type=cpu. With the values being
reported as percentage, collectd sets the type=percent for teh metrics
stored in influx (which grafana retrieves), so the dashboards dodn't
have any data to display. By  removing the type=cpu filter, the CPU
metrics are once again showing in grafana, and will be shown whether the
CPU usage is reported in percentage or the default jiffies.

Signed-off-by: "Emma Foley <efoley@redhat.com>"
Change-Id: I03a6f01677514d888e7805358855947791ab173c